requirements.txt
streamlit
pandas
python-dotenv
# monday-sdk-python
requests